+++ src/objects/components/monster.h
@@ -36,7 +36,7 @@ class _Monster {
 
 		// Attributes
 		_Object *Owner;
-		uint32_t DatabaseID;
+		uint16_t DatabaseID;
 		int ExperienceGiven;
 		int GoldGiven;
 		std::string AI;
